Add json_tree module with all json-based types and implementation.
