+++ components/places/src/storage/bookmarks/public_node.rs
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use super::super::bookmarks::FetchDepth;
+use super::super::bookmarks::json_tree::{fetch_tree, FetchDepth};
 use super::*;
 
 /// This type exists to help when we had protobufs, but now with uniffi it should
