+++ components/places/src/tests.rs
@@ -5,11 +5,14 @@
 use rusqlite::NO_PARAMS;
 use serde_json::Value;
 
+use crate::error::*;
 use crate::{
     db::PlacesDb,
-    storage::bookmarks::{fetch_tree, get_raw_bookmark, insert_tree, BookmarkTreeNode, FetchDepth},
+    storage::bookmarks::get_raw_bookmark,
+    storage::bookmarks::json_tree::{fetch_tree, insert_tree, BookmarkTreeNode, FetchDepth},
     types::BookmarkType,
 };
+
 use sql_support::ConnExt;
 use sync_guid::Guid as SyncGuid;
 use types::Timestamp;
